% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FLXMCpkb}
\alias{FLXMCpkb}
\title{PKB distribution Driver for FlexMix}
\usage{
FLXMCpkb(formula = . ~ .)
}
\arguments{
\item{formula}{A formula.}
}
\value{
Returns an object of class \code{FLXMC}.
}
\description{
This model driver for flexmix implements model-based
clustering of PKB distributions.
}
\examples{
mix <- rbind(rpkb(30, 0.95, c(1, 0, 0)), rpkb(30, 0.9, c(-1, 0, 0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = FLXMCpkb())
}
