% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FLXMCspcauchy}
\alias{FLXMCspcauchy}
\title{Spherical Cauchy routine for flexmix}
\usage{
FLXMCspcauchy(formula = . ~ .)
}
\arguments{
\item{formula}{formula}
}
\value{
Object of type FLXMC for flexmix estimation.
}
\description{
\code{FLXMCspcauchy} offers a flexmix routine for spherical Cauchy distribution.
}
\examples{
mix <- rbind(rPKBD(30, 0.95, c(1,0,0)), rPKBD(30, 0.9, c(-1,0,0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = FLXMCspcauchy())
}
