% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{PKBD_clust}
\alias{PKBD_clust}
\title{PKBD routine for flexmix}
\usage{
PKBD_clust(formula = . ~ .)
}
\arguments{
\item{formula}{formula}
}
\value{
Object of type FLXcomponent for flexmix estimation.
}
\description{
\code{PKBD_clust} offers a flexmix routine for PKBD distribution.
}
