% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{PKBDNN_clust_adam}
\alias{PKBDNN_clust_adam}
\title{PKBD routine using neural networks for flexmix}
\usage{
PKBDNN_clust_adam(
  formula = . ~ .,
  EPOCHS = 100,
  LR = 0.1,
  max_iter = 200,
  line_search_fn = "strong_wolfe",
  free_iter = 5
)

PKBDNN_clust_adam(
  formula = . ~ .,
  EPOCHS = 100,
  LR = 0.1,
  max_iter = 200,
  line_search_fn = "strong_wolfe",
  free_iter = 5
)
}
\arguments{
\item{formula}{formula}

\item{EPOCHS}{number of epochs in the M-step estimation (default: 1)}

\item{LR}{learning rate used in the M-steo estimation (default: 0.5)}

\item{free_iter}{number of initial iterations for which the model in M-step is fully reseted (default: 3)}
}
\value{
Object of type FLXcomponent for flexmix estimation.

Object of type FLXcomponent for flexmix estimation.
}
\description{
\code{PKBDNN_clust_adam} offers a flexmix routine for PKBD distribution using neural networks and adam optimizer.

\code{PKBDNN_clust_adam} offers a flexmix routine for PKBD distribution using neural networks and adam optimizer.
}
