% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBD_sample.R
\name{rpkb}
\alias{rpkb}
\title{Random Sampling from PKB Distributions using ACG or
Projected Saw Distributions}
\usage{
rpkb(n, rho, mu, method = "ACG")
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric giving the concentration parameter.}

\item{mu}{A numeric vector giving the mean direction parameter.}

\item{method}{A character indicating the method to use, \code{"ACG"} for
angular central Gaussian distribution envelopes and \code{"Saw"} for
the use of projected Saw distributions.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from PKB distributions.
}
\examples{
rpkb(10, 0.95, c(1, 0, 0))
}
