% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rPKBD_ACG}
\alias{rPKBD_ACG}
\title{Random Sampling from PKBD Distributions using ACG Envelopes}
\usage{
rPKBD_ACG(n, rho, mu)
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric giving the concentration parameter.}

\item{mu}{A numeric vector giving the mean direction parameter.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from PKBD distributions.
}
\keyword{internal}
