% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBD_sample.R
\name{rPKBD_Saw}
\alias{rPKBD_Saw}
\title{Random Sampling from PKBD Distributions using projected Saw distribution}
\usage{
rPKBD_Saw(n, rho, mu)
}
\arguments{
\item{n}{number of random draws.}

\item{rho}{a numeric giving the concentration parameter.}

\item{mu}{a numeric vector giving the mean direction parameter.}
}
\value{
A vector the generated values.
}
\description{
\code{rPKBD_Saw} generates a random sample from PKBD distributions.
}
\details{
The function generates samples from PKBD using projected Saw distribution.
}
