% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsCauchy}
\alias{rsCauchy}
\title{Random Sampling from spherical Cauchy Distributions}
\usage{
rsCauchy(n, rho, mu)
}
\arguments{
\item{n}{an integer giving the number of samples to draw.}

\item{rho}{a numeric value giving the rho parameters.}

\item{mu}{a numeric vector giving the mu direction parameter.}
}
\value{
A matrix with rows equal to the generated values.
}
\description{
\code{rsCauchy} generates a random sample from a mixture of multivariate Watson distributions.
}
