% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{SCauchy_clust}
\alias{SCauchy_clust}
\title{Spherical Cauchy routine for flexmix}
\usage{
SCauchy_clust(formula = . ~ .)
}
\arguments{
\item{formula}{formula}
}
\value{
Object of type FLXcomponent for flexmix estimation.
}
\description{
\code{SCauchy_clust} offers a flexmix routine for spherical Cauchy distribution.
}
