% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FLXMCpkbd}
\alias{FLXMCpkbd}
\title{PKBD Driver for FlexMix}
\usage{
FLXMCpkbd(formula = . ~ .)
}
\arguments{
\item{formula}{A formula.}
}
\value{
Returns an object of class \code{FLXMC}.
}
\description{
This model driver for flexmix implements model-based
clustering of PKBD distributions.
}
\examples{
mix <- rbind(rPKBD(30, 0.95, c(1, 0, 0)), rPKBD(30, 0.9, c(-1, 0, 0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = FLXMCpkbd())
}
