% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FLXMCspcauchy}
\alias{FLXMCspcauchy}
\title{Spherical Cauchy Driver for FlexMix}
\usage{
FLXMCspcauchy(formula = . ~ .)
}
\arguments{
\item{formula}{A formula.}
}
\value{
Returns an object of class `FLXMC`.
}
\description{
This model driver for flexmix implements model-based
    clustering of spherical Cauchy distributions.
}
\examples{
mix <- rbind(rpkbd(30, 0.95, c(1, 0, 0)), rpkbd(30, 0.9, c(-1, 0, 0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = FLXMCspcauchy())
}
