% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FLXMRspcauchy}
\alias{FLXMRspcauchy}
\title{Spherical Cauchy routine using neural networks for flexmix}
\usage{
FLXMRspcauchy(
  formula = . ~ .,
  EPOCHS = 100,
  LR = 0.1,
  max_iter = 200,
  adam_iter = 5,
  free_iter = adam_iter,
  line_search_fn = "strong_wolfe"
)
}
\arguments{
\item{formula}{formula}

\item{EPOCHS}{number of epochs in the M-step estimation (default: 1)}

\item{LR}{learning rate used in the M-steo estimation (default: 0.5)}

\item{max_iter}{maximum number of iterations of the LBFGS optimizer (default: 200)}

\item{adam_iter}{number of iteration for which the adam optimizer is used before the algorithm switches to L-BFGS (default: 5)}

\item{free_iter}{number of initial iterations for which the model in M-step is fully reseted (default: adam_iter)}

\item{line_search_fn}{method used for line search in LBFGS (default: "strong_wolfe")}
}
\value{
Object of type FLXMC for flexmix estimation.
}
\description{
\code{FLXMRspcauchy} offers a flexmix routine for spherical Cauchy distribution using neural networks and adam optimizer.
}
\examples{
mix <- rbind(rPKBD(30, 0.95, c(1,0,0)), rPKBD(30, 0.9, c(-1,0,0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = FLXMRspcauchy())
}
