% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{PKBD_clust}
\alias{PKBD_clust}
\title{PKBD routine for flexmix}
\usage{
PKBD_clust(formula = . ~ .)
}
\arguments{
\item{formula}{formula}
}
\value{
Object of type FLXMC for flexmix estimation.
}
\description{
\code{PKBD_clust} offers a flexmix routine for PKBD distribution.
}
\examples{
mix <- rbind(rPKBD(30, 0.95, c(1,0,0)), rPKBD(30, 0.9, c(-1,0,0)))
m1 <- flexmix::flexmix(mix ~ 1, k = 2, model = PKBD_clust())
}
