% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{PKBDNN_clust}
\alias{PKBDNN_clust}
\title{PKBD routine using neural networks for flexmix}
\usage{
PKBDNN_clust(
  formula = . ~ .,
  EPOCHS = 1,
  LR = 0.1,
  max_iter = 200,
  line_search_fn = "strong_wolfe",
  free_iter = 3
)
}
\arguments{
\item{formula}{formula}

\item{EPOCHS}{number of epochs in the M-step estimation (default: 1)}

\item{LR}{learning rate used in the M-steo estimation (default: 0.5)}

\item{max_iter}{maximum number of iterations of the LBFGS optimizer (default: 200)}

\item{line_search_fn}{method used for line search in LBFGS (default: "strong_wolfe")}

\item{free_iter}{number of initial iterations for which the model in M-step is fully reseted (default: 3)}
}
\value{
Object of type FLXMC for flexmix estimation.
}
\description{
\code{PKBDNN_clust} offers a flexmix routine for PKBD distribution using neural networks.
}
