% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBD_sample.R
\name{rPKBD}
\alias{rPKBD}
\title{Random Sampling from PKBD Distributions using ACG and projected Saw distribution}
\usage{
rPKBD(n, rho, mu, method = "ACG")
}
\arguments{
\item{n}{number of random draws.}

\item{rho}{a numeric giving the concentration parameter.}

\item{mu}{a numeric vector giving the mean direction parameter.}

\item{method}{method to use, "ACG" for angular central Gaussian distribution envelopes and "Saw" for the use of projected Saw distribution}
}
\value{
A vector the generated values.
}
\description{
\code{rPKBD} generates a random sample from PKBD distributions.
}
\details{
The function generates samples from PKBD using ACG and projected Saw distribution.
}
\examples{
rPKBD(10, 0.95, c(1,0,0))
}
