% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBD_sample.R
\name{rPKBD}
\alias{rPKBD}
\title{Random Sampling from PKBD Distributions using ACG or
    Projected Saw Distributions}
\usage{
rPKBD(n, rho, mu, method = "ACG")
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric giving the concentration parameter.}

\item{mu}{A numeric vector giving the mean direction parameter.}

\item{method}{A character indicating the method to use, `"ACG"` for
angular central Gaussian distribution envelopes and `"Saw"` for
the use of projected Saw distributions.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from PKBD distributions.
}
\examples{
rPKBD(10, 0.95, c(1, 0, 0))
}
