% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rPKBD_ACG}
\alias{rPKBD_ACG}
\title{Random Sampling from PKBD Distributions using ACG Envelopes}
\usage{
rPKBD_ACG(n, lambda, mu)
}
\arguments{
\item{n}{number of random draws.}

\item{lambda}{a numeric giving the concentration parameter.}

\item{mu}{a numeric vector giving the mean direction parameter.}
}
\value{
A vector the generated values.
}
\description{
\code{rPKBD_ACG} generates a random sample from PKBD distributions.
}
\details{
The function generates samples from PKBD using ACG envelopes
}
