% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBD_sample.R
\name{rPKBD_Saw}
\alias{rPKBD_Saw}
\title{Random Sampling from PKB Distributions using Projected Saw Distribution}
\usage{
rPKBD_Saw(n, rho, mu)
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric giving the concentration parameter.}

\item{mu}{A numeric vector giving the mean direction parameter.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from PKB distributions.
}
\keyword{internal}
