% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsCauchy}
\alias{rsCauchy}
\title{Random Sampling from Spherical Cauchy Distributions}
\usage{
rsCauchy(n, rho, mu)
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric value giving the rho parameter.}

\item{mu}{A numeric vector giving the mu direction parameter.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from spherical Cauchy distributions.
}
\examples{
rsCauchy(10, 0.95, c(1,0,0))
}
