% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rscauchy}
\alias{rscauchy}
\title{Random Sampling from Spherical Cauchy Distributions}
\usage{
rscauchy(n, rho, mu)
}
\arguments{
\item{n}{The number of random draws.}

\item{rho}{A numeric value giving the rho parameter.}

\item{mu}{A numeric vector giving the mu direction parameter.}
}
\value{
A matrix with the generated values.
}
\description{
Generates a random sample from spherical Cauchy distributions.
}
\examples{
rscauchy(10, 0.95, c(1, 0, 0))
}
