% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{SCauchy_clust}
\alias{SCauchy_clust}
\title{Spherical Cauchy routine for flexmix}
\usage{
SCauchy_clust(formula = . ~ .)
}
\arguments{
\item{formula}{formula}
}
\value{
Object of type FLXMC for flexmix estimation.
}
\description{
\code{SCauchy_clust} offers a flexmix routine for spherical Cauchy distribution.
}
\examples{
\dontrun{
mix <- rbind(rPKBD(30, 0.95, c(1,0,0)), rPKBD(30, 0.9, c(-1,0,0)))
m1 <- flexmix(mix ~ 1, k = 2, model = SCauchy_clust())
}
}
